+++ packages/api/quantity/lib/quantities/distance/index.ts
@@ -1,5 +1,4 @@
 export * from './distance.js'
-
 export * from './distance.dto.js'
 export * from './distance.column.js'
 export * from './distance-unit.enum.js'