+++ packages/web/design-system/src/ui/dialog/Dialog.vue
@@ -31,7 +31,7 @@ const emit = defineEmits<{
 }>()
 
 const isOpen = defineModel<boolean>('isOpen', {
-  required: true,
+  default: false,
 })
 
 const style = computed<CreateDialogStyle>(() => createDialogStyle({