+++ packages/api/coordinates/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @wisemen/coordinates
 
+## 0.0.25
+
+### Patch Changes
+
+- [#825](https://github.com/wisemen-digital/wisemen-core/pull/825) [`9701b57`](https://github.com/wisemen-digital/wisemen-core/commit/9701b572e17fe10813d592bb80d9440b0159540a) Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - bump dependencies
+
+- Updated dependencies [[`9701b57`](https://github.com/wisemen-digital/wisemen-core/commit/9701b572e17fe10813d592bb80d9440b0159540a)]:
+  - @wisemen/quantity@0.6.1
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [[`959c850`](https://github.com/wisemen-digital/wisemen-core/commit/959c850769865c477309253bde938de94fe2f63f)]:
+  - @wisemen/quantity@0.6.0
+
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [[`8c1e355`](https://github.com/wisemen-digital/wisemen-core/commit/8c1e355e000ec830d383c89261e5edbf5850dee8), [`1308fb4`](https://github.com/wisemen-digital/wisemen-core/commit/1308fb4e12a24d4ddaf4d1a72cecd9ac342d29be)]:
+  - @wisemen/quantity@0.5.0
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [[`b4ada66`](https://github.com/wisemen-digital/wisemen-core/commit/b4ada66c9c9f5c43f1c897d27214854a776493e3)]:
+  - @wisemen/quantity@0.4.1
+
 ## 0.0.21
 
 ### Patch Changes