+++ packages/api/date-time-range/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @wisemen/date-time-range
 
+## 1.0.22
+
+### Patch Changes
+
+- [#825](https://github.com/wisemen-digital/wisemen-core/pull/825) [`9701b57`](https://github.com/wisemen-digital/wisemen-core/commit/9701b572e17fe10813d592bb80d9440b0159540a) Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - bump dependencies
+
+- Updated dependencies [[`9701b57`](https://github.com/wisemen-digital/wisemen-core/commit/9701b572e17fe10813d592bb80d9440b0159540a)]:
+  - @wisemen/pagination@0.1.1
+  - @wisemen/validators@0.0.23
+
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [[`959c850`](https://github.com/wisemen-digital/wisemen-core/commit/959c850769865c477309253bde938de94fe2f63f)]:
+  - @wisemen/pagination@0.1.0
+
 ## 1.0.20
 
 ### Patch Changes