+++ packages/web/generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wisemen/vue-core-generator
 
+## 1.2.3
+
+### Patch Changes
+
+- [#825](https://github.com/wisemen-digital/wisemen-core/pull/825) [`9701b57`](https://github.com/wisemen-digital/wisemen-core/commit/9701b572e17fe10813d592bb80d9440b0159540a) Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - bump dependencies
+
 ## 1.2.2
 
 ### Patch Changes